+++ cmd/shortener/output.json
@@ -1,1 +1,1 @@
-{"short_url":"http://127.0.0.1:8080/948","original_url":"https://golangbyexample2.com/"}
+{"short_url":"http://127.0.0.1:8080/949","original_url":"https://golangbyexample2.com/"}
