+++ cmd/shortener/texts.txt
@@ -946,3 +946,4 @@ https://golangbyexample2.com/
 https://golangbyexample2.com/
 https://golangbyexample2.com/
 https://golangbyexample2.com/
+https://golangbyexample2.com/
